main page + group pages
'index' is the main group page (need to be enhanced)
'Group_chart' is the organization of Dr. Yang's old group members
'Group_members' is the NEW list of the Dr. Yang's group with short bio (need to be enhanced)
'Our_group' is a short explanation of the group and its activities (need to be enhanced)